Add shader code and build script. Road generation and map sectioning

+++ Project/mesh.h
@@ -17,14 +17,14 @@ struct mesh : asset
 	bgfx_vertex_buffer_handle_t vb_handle;
 	u32 vertex_count;
 
-	u16* indices;
+	u32* indices;
 	bgfx_index_buffer_handle_t idb_handle;
 	u32 index_count;
 };
 
 inline u32 pack_vec3_into_u32(vec3 vec);
 
-mesh* mesh_create(char* asset_id, pos_normal_vertex* vertices, u32 vertex_count, u16* indices, u32 index_count);
+mesh* mesh_create(char* asset_id, pos_normal_vertex* vertices, u32 vertex_count, u32* indices, u32 index_count);
 void mesh_destroy(mesh* m);
 
 mesh* load_obj_mesh(char* asset_id, char* filename);
